+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## Package: iqnection-silverstripe-3-modules-payment
+## Package: iqnection-silverstripe-modules-payment
 # Change Log
+
+## [2.0.1]
+- Added base CreditCardPayment class. Requires a processor to carry out the processing
 
 ## [2.0.0]
 - Updated for SilverStripe 4.1